templates/agent-on-sentinel/py/requirements.txt
naylence-agent-sdk>=0.3.27,<0.4.0
python-dotenv>=1.0.1,<2.0.0
